refactor: promote getProcessState to a proper @CamundaMcpTool class
Extract the getProcessState logic from ProcessesToolRepository into a
dedicated ProcessStateTools class that follows the same pattern as other
static tools (IncidentTools, VariableTools, etc.). The framework now
generates the JSON schema from annotations rather than hand-crafting it,
and the tool appears in both /mcp/cluster and /mcp/processes endpoints.

ProcessesToolRepository is simplified to routing only, with no business
logic or service dependencies of its own for the static tool.
